List of functions

This page hosts a list of functions of the Replication_Carleton_et_al_2022.jl package.

Undocumented list of functions

Replication_Carleton_et_al_2022.create_table_1

Replication_Carleton_et_al_2022.delete_Figure_1_estimates

Replication_Carleton_et_al_2022.delete_coefficients

Replication_Carleton_et_al_2022.delete_covar_pop_count

Replication_Carleton_et_al_2022.delete_global_mortality_panel_covariates

Replication_Carleton_et_al_2022.delete_global_mortality_panel_public

Replication_Carleton_et_al_2022.delete_table_1

Replication_Carleton_et_al_2022.load

Replication_Carleton_et_al_2022.load_Figure_1_estimates

Replication_Carleton_et_al_2022.load_coefficients

Replication_Carleton_et_al_2022.load_covar_pop_count

Replication_Carleton_et_al_2022.load_global_mortality_panel_covariates

Replication_Carleton_et_al_2022.test

Documented list of functions

Replication_Carleton_et_al_2022.create_table_1 — Method

The function create_table_1() creates a pdf containing the replication result of the Figure 1.

The pdf file is created within the 0_output folder.

Replication_Carleton_et_al_2022.delete_Figure_1_estimates — Method

The function delete_Figure_1_estimates deletes the file "estimates.csv" from your folder.

Replication_Carleton_et_al_2022.delete_coefficients — Method

The function delete_coefficients deletes the file "coefficients.csv" from your folder.

Replication_Carleton_et_al_2022.delete_covar_pop_count — Method

The function delete_covar_pop_count deletes the file "covarpopcount.dta" from your folder.

Replication_Carleton_et_al_2022.delete_global_mortality_panel_covariates — Method

The function delete_final_data_covariates deletes the file "globalmortalitypanel_covariates.dta" from your folder.

Replication_Carleton_et_al_2022.delete_global_mortality_panel_public — Method

The function delete_global_mortality_panel_public deletes the file "globalmortalitypanel_public.dta" from your folder.

Replication_Carleton_et_al_2022.delete_table_1 — Method

The functio 'deletetable1()' deletes the pdf of the table 1 in the output folder.

Replication_Carleton_et_al_2022.load — Method

The function load ensures that all the required data is in the local folder of the user. It should be run before the run function of the package.

Replication_Carleton_et_al_2022.load_Figure_1_estimates — Method

The function loading_Figure_1_estimates ensures that the file "estimates.csv" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_coefficients — Method

The function load_coefficients ensures that the file "coefficients.csv" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_covar_pop_count — Method

The function load_covar_pop_count ensures that the file "covarpopcount.dta" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_global_mortality_panel_covariates — Method

This function did not get tested!!!

The function load_final_data_covariates ensures that the file "globalmortalitypanel_covariates" is in your input folder. Since the data weights 2 Gigabytes, be sure to have a good internet connection.

Replication_Carleton_et_al_2022.test — Method

The function test is a test to check that the Replication_Carleton_et_al_2022 module is well loaded.