Technical documentation

This page hosts the technical documentation of the Replication_Carleton_et_al_2022.jl package.

Installation

The installation of the package relies on the built-in Pkg package, by specifying the url of its GitHub repository.

using Pkg
Pkg.activate(".")
Pkg.add(url = "https://github.com/Paulogcd/Replication_Carleton_et_al_2022.jl")
using Replication_Carleton_et_al_2022 

Main functions

The package contains two main functions:

The load() function, that ensures that all files required for the data analysis are in your folder

The run() function, that generates all the results we managed to replicate in the an output folder

# Ensures that an adequate folder structure is generated, and download the required files in those folders
Replication_Carleton_et_al_2022.load()

# Generate the exhibits we managed to replicate from the original article
Replication_Carleton_et_al_2022.run()

Undocumented list of functions

Replication_Carleton_et_al_2022.create_figure_2

Replication_Carleton_et_al_2022.create_folder_setup

Replication_Carleton_et_al_2022.create_table_1

Replication_Carleton_et_al_2022.delete_Figure_1_estimates

Replication_Carleton_et_al_2022.delete_coefficients

Replication_Carleton_et_al_2022.delete_covar_pop_count

Replication_Carleton_et_al_2022.delete_global_mortality_panel_covariates

Replication_Carleton_et_al_2022.delete_global_mortality_panel_public

Replication_Carleton_et_al_2022.delete_table_1

Replication_Carleton_et_al_2022.load

Replication_Carleton_et_al_2022.load_Figure_1_estimates

Replication_Carleton_et_al_2022.load_coefficients

Replication_Carleton_et_al_2022.load_covar_pop_count

Replication_Carleton_et_al_2022.load_global_mortality_panel_covariates

Replication_Carleton_et_al_2022.run

Replication_Carleton_et_al_2022.test

Documented list of functions

Replication_Carleton_et_al_2022.create_figure_2 — Method

The function "createfigure2()" generates our attempt of replication of the figure 2 in the "output()" folder. It creates two png files, respectively "Figure22015" and "Figure22100".

Running a @time evaluation on this function, a MacBook Pro M4 (16Gb of RAM) obtains: 13.516709 seconds (83.77 M allocations: 5.982 GiB, 20.97% gc time)

Replication_Carleton_et_al_2022.create_folder_setup — Method

The function "createfoldersetup" creates the folders "0input" (and its subfolders) and "0output", that are necessary for the other functions to run.

Replication_Carleton_et_al_2022.create_table_1 — Method

The function create_table_1() creates a pdf containing the replication result of the Figure 1.

The pdf file is created within the 0_output folder.

Replication_Carleton_et_al_2022.delete_Figure_1_estimates — Method

The function delete_Figure_1_estimates deletes the file "estimates.csv" from your folder.

Replication_Carleton_et_al_2022.delete_coefficients — Method

The function delete_coefficients deletes the file "coefficients.csv" from your folder.

Replication_Carleton_et_al_2022.delete_covar_pop_count — Method

The function delete_covar_pop_count deletes the file "covarpopcount.dta" from your folder.

Replication_Carleton_et_al_2022.delete_global_mortality_panel_covariates — Method

The function delete_final_data_covariates deletes the file "globalmortalitypanel_covariates.dta" from your folder.

Replication_Carleton_et_al_2022.delete_global_mortality_panel_public — Method

The function delete_global_mortality_panel_public deletes the file "globalmortalitypanel_public.dta" from your folder.

Replication_Carleton_et_al_2022.delete_table_1 — Method

The functio 'deletetable1()' deletes the pdf of the table 1 in the output folder.

Replication_Carleton_et_al_2022.load — Method

The function load ensures that all the required data is in the local folder of the user. It should be run before the run function of the package.

Replication_Carleton_et_al_2022.load_Figure_1_estimates — Method

The function loading_Figure_1_estimates ensures that the file "estimates.csv" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_coefficients — Method

The function load_coefficients ensures that the file "coefficients.csv" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_covar_pop_count — Method

The function load_covar_pop_count ensures that the file "covarpopcount.dta" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_global_mortality_panel_covariates — Method

This function did not get tested!!!

The function load_final_data_covariates ensures that the file "globalmortalitypanel_covariates" is in your input folder. Since the data weights 2 Gigabytes, be sure to have a good internet connection.

Replication_Carleton_et_al_2022.run — Method

The function "run()" generates all the results we were able to replicate in an "output" folder. It also ensures the required data for the replication is available locally. It may download the files if you do not have them correctly set up.

Replication_Carleton_et_al_2022.test — Method

The function test is a test to check that the Replication_Carleton_et_al_2022 module is well loaded.