Technical documentation

This page hosts the technical documentation of the Replication_Carleton_et_al_2022.jl package.

Installation

This package requires Julia (version 1.11.4 was used) to run. Once you are located in a chosen directory, you can launch Julia on your Terminal by entering:

julia

Then, once Julia has started, we are going to need the Pkg module to load the current replication package. We are going to create a local environment, and add to this environment the current replication package. Finally, we are going to load it.

# Load the Pkg package:
using Pkg                                   
# Activate an environment at the current directory:
Pkg.activate(".")                           
# Download the current replication package:
Pkg.add(url = "https://github.com/Paulogcd/Replication_Carleton_et_al_2022.jl")         
# Load the replication package:
using Replication_Carleton_et_al_2022 

Now, the package is loaded in your local environment. You can test that it is loaded by using the test() function:

# To test the loading of the package:
Replication_Carleton_et_al_2022.test()

This function should print a message indicating that the package is well-loaded.

Required files

To maintain the portability of the package, we did not include the large files required to perform the computations directly in it.

This is why we need to download these files via the load() function. It will first create two main folders, 0_input and 0_output, and will then download the files in their corresponding subfolders.

Caution: make sure that you have 5.3 Gigabytes of space available for the package.

# To download the required files:
Replication_Carleton_et_al_2022.load()

Replication results

Once the required files are downloaded, we can generate the replication results, with the run() function:

# To generate the replication results:
Replication_Carleton_et_al_2022.run()

This function will perform the replication computations and will store the results in the '0_output' folder. The replication results corredspond to:

Table 1: Descriptive statistics of the initial dataset the authors use.

Figure 1: Plots of function describing the mortality-temperature relationship, based on the theoretical framework of the authors.

Figure 2: A heatmap showing the predicted mortality effect.

Delete required files

Due to the large size of the package files, we also included a function to manage their deletion. If you wish to keep your directory, but without the 0_input and 0_output folders, you can delete them via the following function:

Replication_Carleton_et_al_2022.delete_folder_setup()

Be careful. This is irreversible.

Undocumented list of functions of the package

Replication_Carleton_et_al_2022.create_figure_2

Replication_Carleton_et_al_2022.create_folder_setup

Replication_Carleton_et_al_2022.create_table_1

Replication_Carleton_et_al_2022.delete_Figure_1_estimates

Replication_Carleton_et_al_2022.delete_coefficients

Replication_Carleton_et_al_2022.delete_covar_pop_count

Replication_Carleton_et_al_2022.delete_global_mortality_panel_covariates

Replication_Carleton_et_al_2022.delete_global_mortality_panel_public

Replication_Carleton_et_al_2022.delete_mortality_allpreds_filtered

Replication_Carleton_et_al_2022.load

Replication_Carleton_et_al_2022.load_Figure_1_estimates

Replication_Carleton_et_al_2022.load_coefficients

Replication_Carleton_et_al_2022.load_covar_pop_count

Replication_Carleton_et_al_2022.load_global_mortality_panel_covariates

Replication_Carleton_et_al_2022.load_mortality_allpreds_filtered

Replication_Carleton_et_al_2022.old_create_table_1

Replication_Carleton_et_al_2022.run

Replication_Carleton_et_al_2022.test

Documented list of functions of the package

Replication_Carleton_et_al_2022.create_figure_2 — Method

The function "createfigure2()" generates our attempt of replication of the figure 2 in the "output()" folder. It creates two png files, respectively "Figure22015" and "Figure22100".

Running a @time evaluation on this function, a MacBook Pro M4 (16Gb of RAM) obtains: 13.516709 seconds (83.77 M allocations: 5.982 GiB, 20.97% gc time)

Replication_Carleton_et_al_2022.create_folder_setup — Method

The function "createfoldersetup" creates the folders "0input" (and its subfolders) and "0output", that are necessary for the other functions to run.

Replication_Carleton_et_al_2022.create_table_1 — Method

creates a html file containing the replication result of the Figure 1.

The html file is created within the 0_output folder.

If you want to generate the Table 1 in a pdf file, you should use the oldcreatetable_1() function. You should however ensure that you have latex installed to do so.

Replication_Carleton_et_al_2022.delete_Figure_1_estimates — Method

The function delete_Figure_1_estimates deletes the file "estimates.csv" from your folder.

Replication_Carleton_et_al_2022.delete_coefficients — Method

The function delete_coefficients deletes the file "coefficients.csv" from your folder.

Replication_Carleton_et_al_2022.delete_covar_pop_count — Method

The function delete_covar_pop_count deletes the file "covarpopcount.dta" from your folder.

Replication_Carleton_et_al_2022.delete_global_mortality_panel_covariates — Method

The function delete_final_data_covariates deletes the file "globalmortalitypanel_covariates.dta" from your folder.

Replication_Carleton_et_al_2022.delete_global_mortality_panel_public — Method

The function delete_global_mortality_panel_public deletes the file "globalmortalitypanel_public.dta" from your folder.

Replication_Carleton_et_al_2022.delete_mortality_allpreds_filtered — Method

The function delete_mortality_allpreds_filtered deletes the file "globalmortalitypanel_covariates.csv" from your folder.

Replication_Carleton_et_al_2022.load — Method

The function load ensures that all the required data is in the local folder of the user. It should be run before the run function of the package.

Replication_Carleton_et_al_2022.load_Figure_1_estimates — Method

The function loading_Figure_1_estimates ensures that the file "estimates.csv" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_coefficients — Method

The function load_coefficients ensures that the file "coefficients.csv" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_covar_pop_count — Method

The function load_covar_pop_count ensures that the file "covarpopcount.dta" is in your input folder. It is necessary for the replication of the Figure 1 of the article.

Replication_Carleton_et_al_2022.load_global_mortality_panel_covariates — Method

The function load_final_data_covariates ensures that the file "globalmortalitypanel_covariates" is in your input folder. Since the data weights 2 Gigabytes, be sure to have a good internet connection.

Replication_Carleton_et_al_2022.load_mortality_allpreds_filtered — Method

The function load_final_data_covariates ensures that the file "globalmortalitypanel_covariates" is in your input folder. Since the data weights 3 Gigabytes, be sure to have a good internet connection.

Replication_Carleton_et_al_2022.old_create_table_1 — Method

The function old_create_table_1() can only be run if you have a LaTeX distribution installed on your machine. Use create_table_1() instead, to produce an HTML version of the replication result of the table 1.

Replication_Carleton_et_al_2022.run — Method

The function "run()" generates all the results we were able to replicate in an "output" folder. It also ensures the required data for the replication is available locally. It may download the files if you do not have them correctly set up.

Replication_Carleton_et_al_2022.test — Method

The function test is a test to check that the Replication_Carleton_et_al_2022 module is well loaded.